Move crawlers to their own folder
