+++ src/crawlers/run_statistics.js
@@ -1,0 +1,3 @@
+class RunStatistics {
+
+}
